Store image-sized correction/lookup maps and NXmx float reads as float
Continue the image-as-float effort in the foundational read path.

- imageset.h: ExternalLookup gain/pedestal/dx/dy and get_gain/get_pedestal/
  get_corrected_data store image-sized maps at float precision. The python API
  and pickle remain double via float<->double conversion in imageset_ext.cc, so
  the on-disk format is unchanged and all consumers (serialize, experiment_list,
  dials.import, the OffsetPxMmStrategy geometry path) keep working.
- nexus/__init__.py _dataset_as_flex: read all float datasets at single
  precision. This is a no-op versus prior behavior (the raw-data cache stores
  float32 regardless) and harmonizes with the legacy reader, which already
  downgrades float64 -> float32.
- tests/nexus/test_build_dxtbx_models.py: update the float64 dataset test to
  assert flex.float.

+++ tests/nexus/test_build_dxtbx_models.py
@@ -803,6 +803,9 @@ def test_dataset_as_flex_float():
 
 
 def test_dataset_as_flex_double():
+    # Double-precision (and any >4-byte) float datasets are now read at single
+    # precision: DIALS integration is float and the raw-data cache stores float32
+    # regardless, so reading float64 would only create a transient double buffer.
     slices = ()
     np_double_types = (
         np.float64,
@@ -814,7 +817,7 @@ def test_dataset_as_flex_double():
         for i, dtype in enumerate(np_double_types):
             d = g.create_dataset(f"double-{i}", data=np.array([0, 1], dtype=dtype))
             flex_a = dxtbx.nexus._dataset_as_flex(d, slices)
-            assert isinstance(flex_a, flex.double)
+            assert isinstance(flex_a, flex.float)
             assert flex_a.all() == d.shape
             assert list(flex_a) == list(d)
